update image so it uses size_t instead of int for dimensions
git-svn-id: file:///home/roberto/Git/svn@6512 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: dd78c549bfb13e46b3666f324405c2922b9a92f7

+++ trunk/xmipp/libraries/data/image.cpp
@@ -86,9 +86,9 @@ void Image< std::complex< double > >::castPage2Datatype(std::complex<double> * s
 }
 
 // Get size of datatype
-unsigned long  gettypesize(DataType type)
+size_t  gettypesize(DataType type)
 {
-    unsigned long   size;
+    size_t   size;
 
     switch ( type ) {
         case UChar: case SChar:  size = sizeof(char); break;
